Switch to internal files directory for storing books (#129)
This fixes the issue of books failing to open randomly over time due to the app losing access to EPUB files it created. This loss of access can occur for various reasons, such as uninstalling and reinstalling the app, MediaStore reindexing, or another app gaining access to the download directory, resulting in write/read access loss. These issues are difficult to reproduce. Two possible fixes I can think of are requesting access to all files, as file managers do, or switching to the app-specific internal files directory. The former option, of course, makes no sense, so we're going with the latter. If you need access to the EPUB file of a certain book, you can always swipe right to save or share it anywhere you like.